refactored and reconfigured functions that query and process data for frontend
